+++ backend/agentd/agentd.go
@@ -1,6 +1,7 @@
 package agentd
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"net/http"
@@ -86,7 +87,7 @@ func (a *Agentd) Start() error {
 
 // Stop Agentd.
 func (a *Agentd) Stop() error {
-	if err := a.httpServer.Shutdown(nil); err != nil {
+	if err := a.httpServer.Shutdown(context.TODO()); err != nil {
 		// failure/timeout shutting down the server gracefully
 		logger.Error("failed to shutdown http server gracefully - forcing shutdown")
 		if closeErr := a.httpServer.Close(); closeErr != nil {