+++ backend/apid/middlewares/allowlist.go
@@ -31,6 +31,5 @@ func (m AllowList) Then(next http.Handler) http.Handler {
 		}
 
 		next.ServeHTTP(w, r)
-		return
 	})
 }