+++ build.sh
@@ -38,7 +38,7 @@ install_deps () {
     go get gopkg.in/alecthomas/gometalinter.v1
     go get github.com/gordonklaus/ineffassign
     go get github.com/jgautheron/goconst/cmd/goconst
-    go get github.com/kisielk/errcheck
+    go get honnef.co/go/tools/cmd/megacheck
     go get github.com/golang/lint/golint
     go get github.com/UnnoTed/fileb0x
     install_golang_dep
@@ -160,7 +160,7 @@ linter_commands () {
         exit 1
     fi
 
-    errcheck $(go list ./... | grep -v dashboardd | grep -v agent/assetmanager | grep -v scripts)
+    megacheck $(go list ./... | grep -v dashboardd | grep -v agent/assetmanager | grep -v scripts)
     if [ $? -ne 0 ]; then
         echo "Linting failed..."
         exit 1