+++ types/check.go
@@ -196,7 +196,7 @@ func (c *CheckConfig) Validate() error {
 
 // Validate returns an error if the ProxyRequests does not pass validation tests
 func (p *ProxyRequests) Validate() error {
-	if (p.SplayCoverage < 0) || (p.SplayCoverage > 100) {
+	if p.SplayCoverage > 100 {
 		return errors.New("proxy request splay coverage must be between 0 and 100")
 	}