+++ CHANGELOG.md
@@ -40,6 +40,7 @@ subcommand.
 - Agent backend URLs without a port specified will now default to port 8081.
 - Travis encrypted variables have been updated to work with travis-ci.org
 - Upgraded all builds to use Go 1.10.
+- Use megacheck instead of errcheck.
 
 ### Fixed
 - Fixed a bug in time.InWindow that in some cases would cause subdued checks to